fix: skip docker installation script if already installed

+++ scripts/apps.install.zsh
@@ -67,23 +67,30 @@ if [[ $SYS_NAME = "mac" ]]; then
     # docker
     if [[ ${DOTFILES_APPS[docker]} = "true" ]]; then
 
-        local _sys_archt=${SYS_ARCHT:-$(uname -m)}
+        if type docker >"/dev/null"; then
 
-        # ----- docker
+            echo_app_installation_message 'docker' 'skip'
 
-        # download
-        curl -O "https://desktop.docker.com/mac/main/$_sys_archt/Docker.dmg?utm_source=docker&utm_medium=webreferral&utm_campaign=docs-driven-download-mac-$_sys_archt"
+        else
 
-        # install
-        sudo hdiutil attach Docker.dmg
-        sudo /Volumes/Docker/Docker.app/Contents/MacOS/install
+            echo_app_installation_message 'docker' 'start'
 
-        # clean up
-        sudo hdiutil detach /Volumes/Docker
-        rm Docker.dmg
+            # ----- docker
 
-        # ----- docker-buildx
-        brew install docker-buildx
+            # download
+            curl -O "https://desktop.docker.com/mac/main/$SYS_ARCHT/Docker.dmg?utm_source=docker&utm_medium=webreferral&utm_campaign=docs-driven-download-mac-$SYS_ARCHT"
+
+            # install
+            sudo hdiutil attach Docker.dmg
+            sudo /Volumes/Docker/Docker.app/Contents/MacOS/install
+
+            # clean up
+            sudo hdiutil detach /Volumes/Docker
+            rm Docker.dmg
+
+            # ----- docker-buildx
+            brew install docker-buildx
+        fi
     fi
 
     # elasticsearch
